+++ src/AngleSharp.Css.Tests/Declarations/CssTextProperty.cs
@@ -690,6 +690,39 @@ namespace AngleSharp.Css.Tests.Declarations
 			Assert.IsFalse(property.IsInherited);
 			Assert.IsFalse(property.IsImportant);
 			Assert.IsFalse(property.HasValue);
-		}
-	}
+        }
+
+        [Test]
+        public void CssTextAlignLegalStart_Issue151()
+        {
+            var snippet = "text-align:start";
+            var property = ParseDeclaration(snippet);
+            Assert.AreEqual("text-align", property.Name);
+            Assert.IsTrue(property.HasValue);
+            Assert.IsFalse(property.IsImportant);
+            Assert.IsFalse(property.IsInherited);
+            Assert.AreEqual("start", property.Value);
+        }
+
+        [Test]
+        public void CssTextAlignLegalJustifyAll_Issue151()
+        {
+            var snippet = "text-align:justify-all";
+            var property = ParseDeclaration(snippet);
+            Assert.AreEqual("text-align", property.Name);
+            Assert.IsTrue(property.HasValue);
+            Assert.IsFalse(property.IsImportant);
+            Assert.IsFalse(property.IsInherited);
+            Assert.AreEqual("justify-all", property.Value);
+        }
+
+        [Test]
+        public void CssTextAlignIllegalJustifyNone_Issue151()
+        {
+            var snippet = "text-align:justify-none";
+            var property = ParseDeclaration(snippet);
+            Assert.AreEqual("text-align", property.Name);
+            Assert.IsFalse(property.HasValue);
+        }
+    }
 }